Revert "updated admin functionalities"
