+++ Water-Transport-System/src/main/java/Team/Gamma/Water_Transport_System/Service/ShipDetailsService.java
@@ -4,7 +4,9 @@ import java.util.List;
 
 public interface ShipDetailsService {
 
-
+    public String addShip(ShipDetail ship);
+    public String editShip(ShipDetail ship);
+    public String deleteShip(Long shipId);
     public ShipDetail getShip(Long adminId);
     public List<ShipDetail> getAllShips();
     List<ShipDetail> searchCruise(String source, String destination);
